restructure package to be multiple files instead of just one main script
Fixes #4
