+++ jmu_openconnect/auth.py
@@ -1,13 +1,11 @@
-#!/usr/bin/env python3
+"""Module for user authenticate with selenium.
 
-import argparse
-import getpass
+This module contains utilities related to authenticating the user with selenium,
+such as the Browser enum as well as the function for fetching the DSID cookie.
+"""
+
 import json
 import logging
-import os
-import shutil
-import subprocess
-import sys
 import time
 from enum import Enum
 from typing import Optional
@@ -23,6 +21,8 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as EC  # noqa: N812
 from selenium.webdriver.support.wait import WebDriverWait
 
+from jmu_openconnect.exceptions import MissingDSIDError, TimedOutError
+
 
 class Browser(Enum):
 	"""Specify a browser to use"""
@@ -30,42 +30,6 @@ class Browser(Enum):
 	FIREFOX = {'driver': webdriver.Firefox, 'options': webdriver.FirefoxOptions}
 	CHROME = {'driver': webdriver.Chrome, 'options': webdriver.ChromeOptions}
 	EDGE = {'driver': webdriver.Edge, 'options': webdriver.EdgeOptions}
-
-
-class MissingDSIDError(Exception):
-	"""Raised if DSID cooke was not found."""
-
-	def __init__(
-		self,
-		msg='DSID Cookie was not found',
-		*args,
-		**kwargs,
-	):
-		super().__init__(msg, *args, **kwargs)
-
-
-class TimedOutError(Exception):
-	"""Raised if the webdriver timed out while waiting for authentication."""
-
-	def __init__(
-		self,
-		msg='webdriver timed out while waiting for authentication',
-		*args,
-		**kwargs,
-	):
-		super().__init__(msg, *args, **kwargs)
-
-
-class MissingOpenConnectError(Exception):
-	"""If openconnect could not be found on the PATH."""
-
-	def __init__(
-		self,
-		msg="openconnect binary could not be found. make sure it's installed and on your PATH",
-		*args,
-		**kwargs,
-	):
-		super().__init__(msg, *args, **kwargs)
 
 
 def fetch_latest_browser_version(browser: Browser) -> str:
@@ -257,175 +221,3 @@ def get_dsid_cookie(
 	driver.quit()
 
 	return dsid_cookie['value']
-
-
-def start_openconnect(dsid_cookie: str) -> int:
-	"""Start openconnect with the specified DSID cookie.
-
-	Args:
-		dsid_cookie (str): The DSID cookie to authenticate with
-
-	Raises:
-		MissingOpenConnectError: If openconnect couldn't be found on the PATH
-		PermissionError: If the user isn't root and sudo/doas couldn't be found
-
-	Returns:
-		int: the openconnect return code
-	"""
-	logging.info('Starting openconnect')
-
-	logging.debug('Checking if openconnect is installed')
-	if shutil.which('openconnect') is None:
-		raise MissingOpenConnectError
-
-	as_root = next(([prog] for prog in ('doas', 'sudo') if shutil.which(prog)), [])
-	logging.debug(f'Root program identified: {as_root}')
-
-	# check if the script is running as root or if sudo/doas were not found
-	# os.geteuid() will be 0 if root
-	if os.geteuid() and not as_root:
-		raise PermissionError('sudo/doas were not found')
-
-	oc_command = as_root + [
-		'openconnect',
-		'--protocol',
-		'pulse',
-		'--cookie',
-		dsid_cookie,
-		'https://vpn.jmu.edu',
-	]
-
-	logging.debug(f'OC Command: {oc_command}')
-
-	# we use subprocess.call to prompt the user for root if needed, and return the return code
-	return subprocess.call(oc_command)
-
-
-def main():
-	logging.basicConfig(
-		format='%(asctime)s | %(levelname)s | %(module)s:%(module)s:%(lineno)d - %(message)s',
-		level=os.environ.get('LOGLEVEL', 'INFO').upper(),
-	)
-
-	parser = argparse.ArgumentParser(
-		description='OpenConnect helper for logging into the JMU Ivanti VPN through Duo'
-	)
-
-	parser.add_argument(
-		'--browser',
-		'-b',
-		choices=['firefox', 'chrome', 'edge'],
-		default='firefox',
-		nargs='?',
-		help='Which web browser to use in Duo authentication. Default firefox',
-	)
-
-	parser.add_argument(
-		'--browser-binary',
-		'-B',
-		help="Specify a path to your browser's binary. Useful if you use a derivative of a supported browser, like WaterFox.",
-	)
-
-	parser.add_argument(
-		'--username',
-		'-u',
-		default='',
-		help='Automatically type in a username',
-	)
-
-	parser.add_argument(
-		'--password',
-		'-p',
-		default='',
-		help='Automatically type in a password',
-	)
-
-	parser.add_argument(
-		'-P',
-		'--prompt-password',
-		action='store_true',
-		help='Prompt for the password without echoing as to not show it in your command history',
-	)
-
-	parser.add_argument(
-		'--timeout',
-		type=int,
-		default=300,
-		help='Number of seconds it takes before the webdriver times out waiting for authentication. Default 300 seconds',
-	)
-
-	parser.add_argument(
-		'--debug-auth-error',
-		action='store_true',
-		help='Pause for 10 seconds after authentication. Useful for debugging errors',
-	)
-
-	parser.add_argument(
-		'-A',
-		'--only-authenticate',
-		action='store_true',
-		help="Only authenticate and don't start openconnect. Prints the DSID cookie to STDOUT",
-	)
-
-	args = parser.parse_args()
-
-	# check if we need to prompt the user for a password
-	password = args.password
-	if args.prompt_password:
-		password = getpass.getpass('Password: ')
-
-	# assign the browser to use. argparse guarentees that it will be one of these three
-	browser = {
-		'firefox': Browser.FIREFOX,
-		'chrome': Browser.CHROME,
-		'edge': Browser.EDGE,
-	}[args.browser]
-
-	if not args.only_authenticate:
-		# check for openconnect before making the user authenticate
-		logging.debug('Checking if openconnect is installed')
-		if shutil.which('openconnect') is None:
-			print('openconnect was not found on your PATH. Please ensure that openconnect is installed.')
-			sys.exit(1)
-
-	try:
-		dsid_cookie = get_dsid_cookie(
-			username=args.username,
-			password=password,
-			browser=browser,
-			browser_binary=args.browser_binary,
-			webdriver_timeout=args.timeout,
-			debug_auth_error=args.debug_auth_error,
-		)
-	except TimedOutError:
-		print('webdriver timed out while waiting for authentication')
-		sys.exit(1)
-	except MissingDSIDError:
-		print(
-			'DSID Cookie was not found. Ensure that you are not logged in multiple times. See README for details.'
-		)
-		sys.exit(1)
-
-	if args.only_authenticate:
-		print(dsid_cookie)
-		sys.exit(0)
-
-	exit_code = 0
-	try:
-		exit_code = start_openconnect(dsid_cookie=dsid_cookie)
-	except MissingOpenConnectError:
-		print(
-			'openconnect was not found on your PATH. Please ensure that openconnect is installed.'
-		)
-		exit_code = 1
-	except PermissionError:
-		print(
-			'sudo was not found on your PATH. Please install sudo or run this script as root.'
-		)
-		exit_code = 1
-
-	sys.exit(exit_code)
-
-
-if __name__ == '__main__':
-	main()
